+++ docs-demo/components/CheckItem.vue
@@ -8,10 +8,17 @@ const props = withDefaults(
     },
 );
 const modelValue = defineModel<boolean>();
+const emit = defineEmits<{
+    (e: 'change', value: boolean): void;
+}>();
+function handleChange(e: Event) {
+    const target = e.target as HTMLInputElement;
+    emit('change', target.checked);
+}
 </script>
 <template>
     <label>
-        <input v-model="modelValue" type="checkbox" />
+        <input v-model="modelValue" type="checkbox" @change="handleChange" />
         <span>{{ props.text }}</span>
     </label>
 </template>